feat: add async support and api call
