+++ cmd/start.go
@@ -2,7 +2,8 @@ package cmd
 
 import (
 	"github.com/spf13/cobra"
-	"orbital/node"
+	"orbital/internal/auth"
+	"orbital/orbital"
 )
 
 var startCmd = &cobra.Command{
@@ -11,7 +12,9 @@ var startCmd = &cobra.Command{
 	RunE: func(cmd *cobra.Command, args []string) error {
 		cmdHeader("start")
 
-		orbitalNode := node.New()
+		apiSrv := setupAPIServer()
+
+		orbitalNode := orbital.New(apiSrv)
 		if err := orbitalNode.Start(); err != nil {
 			return err
 		}
@@ -21,4 +24,15 @@ var startCmd = &cobra.Command{
 
 func init() {
 	rootCmd.AddCommand(startCmd)
+}
+
+func setupAPIServer() *orbital.Server {
+	apiSrv := orbital.NewServer()
+
+	helloService := auth.NewService()
+
+	// Register all services to apiServer
+	auth.RegisterHelloServiceServer(apiSrv, helloService)
+
+	return apiSrv
 }
