+++ dashboard/wasm/api/api.go
@@ -6,6 +6,7 @@ import (
 	"fmt"
 	"io"
 	"net/http"
+	"strings"
 	"time"
 )
 
@@ -19,7 +20,7 @@ func NewAPI(basePath string) *API {
 		defaultHeaders: map[string]string{
 			"Content-Type":    "application/json",
 			"Accept":          "application/json",
-			"Accept-Encoding": "gzip, br",
+			"Accept-Encoding": "gzip",
 		},
 		urlPath: basePath,
 	}
@@ -27,6 +28,8 @@ func NewAPI(basePath string) *API {
 
 func (api *API) Do(data []byte, headers map[string]string) ([]byte, error) {
 	req, err := http.NewRequest(http.MethodPost, api.urlPath, bytes.NewBuffer(data))
+	fmt.Println("Request URL:", api.urlPath)
+
 	if err != nil {
 		return nil, err
 	}
@@ -35,30 +38,26 @@ func (api *API) Do(data []byte, headers map[string]string) ([]byte, error) {
 		headers = make(map[string]string)
 	}
 
-	reqHeaders := api.mergeHeaders(headers)
+	api.mergeHeaders(req, headers)
 	client := &http.Client{
-		Timeout: 5 * time.Second,
+		Transport: &http.Transport{
+			ForceAttemptHTTP2: false,
+		},
+		Timeout: time.Second * 5,
 	}
-
-	for k, v := range reqHeaders {
-		req.Header.Set(k, v)
-	}
-
 	response, err := client.Do(req)
 	if err != nil {
+		fmt.Printf("ERR1: %s\n", err)
 		return nil, err
 	}
-
 	defer response.Body.Close()
 
-	var bodyReader io.ReadCloser = response.Body
-	if response.Header.Get("Content-Encoding") == "gzip" {
-
+	var bodyReader = response.Body
+	if strings.Contains(response.Header.Get("Content-Encoding"), "gzip") {
 		bodyReader, err = gzip.NewReader(response.Body)
 		if err != nil {
 			return nil, fmt.Errorf("failed to decompress response body: %w", err)
 		}
-
 		defer bodyReader.Close()
 	}
 
@@ -71,18 +70,17 @@ func (api *API) Do(data []byte, headers map[string]string) ([]byte, error) {
 		return nil, fmt.Errorf("unexpected response status code: %d, body: %s", response.StatusCode, body)
 	}
 
+	fmt.Printf("res: %s\n", string(body))
 	return body, nil
 }
 
-func (api *API) mergeHeaders(extraHeaders map[string]string) map[string]string {
-	headersMerged := make(map[string]string)
+func (api *API) mergeHeaders(r *http.Request, extraHeaders map[string]string) {
 	for key, value := range api.defaultHeaders {
-		headersMerged[key] = value
+		r.Header.Set(key, value)
 	}
 
 	for key, value := range extraHeaders {
-		headersMerged[key] = value
+		r.Header.Set(key, value)
 	}
 
-	return headersMerged
 }
