+++ dashboard/wasm/components/login.go
@@ -1,7 +1,9 @@
 package components
 
 import (
+	"encoding/json"
 	"fmt"
+	"orbital/dashboard/wasm/api"
 	"orbital/dashboard/wasm/dom"
 	"orbital/dashboard/wasm/events"
 	"orbital/dashboard/wasm/storage"
@@ -14,6 +16,7 @@ type LoginComponentDI struct {
 }
 
 type LoginComponent struct {
+	async     *api.Async
 	secretKey string
 	tplDir    string
 	events    *events.Event
@@ -50,12 +53,48 @@ func (c *LoginComponent) Show() {
 
 func (c *LoginComponent) uiLoginAction(this js.Value, args []js.Value) interface{} {
 	input := dom.DocQuerySelectorValue("#public-key", "value")
-	fmt.Println("Validate:", input)
+
+	req := map[string]string{
+		"publicKey": input.String(),
+	}
+
+	reqBin, err := json.Marshal(req)
+	if err != nil {
+		dom.PrintToConsole(fmt.Sprintf("Error serializing public key: %v", err))
+		return nil
+	}
+
+	resultChan := c.async.RunWithResult(func() (interface{}, error) {
+		client := api.NewAPI("/rpc/AuthService/Auth")
+		res, err := client.Do(reqBin, nil)
+		if err != nil {
+			dom.PrintToConsole(fmt.Sprintf("Error calling HelloService: %v", err))
+			return nil, err
+		}
+
+		resMap := make(map[string]interface{})
+		_ = json.Unmarshal(res, &resMap)
+
+		fmt.Println("Response:", resMap)
+		return resMap, nil
+	})
+
+	go func() {
+		result := <-resultChan
+		if result.Err != nil {
+			fmt.Printf("Error validating public key: %v\n", result.Err)
+			return
+		}
+
+		fmt.Printf("Validation successful, response: %s\n", result.Value)
+	}()
+
 	return nil
 }
 
 func NewLoginComponents(di LoginComponentDI) {
 	c := &LoginComponent{
+		async:  api.NewAsync(),
 		events: di.Events,
 		store:  di.Storage,
 		tplDir: "login",
